worldcompass/requirements.txt
hpsv3==1.0.0
torch==2.7.1
torchvision==0.22.1
torchdata==0.11.0
triton==3.3.1
tensorboard==2.20.0
diffusers==0.35.1
huggingface_hub==0.35.0
safetensors==0.4.5
peft==0.17.1
accelerate==1.10.1
numpy==1.26.4
pandas==2.3.2
scipy==1.15.1
pyarrow==21.0.0
Pillow==12.0.0
pillow_heif==1.1.1
imageio==2.37.0
moviepy==1.0.3
opencv_python==4.11.0.86
trimesh==4.6.2
plyfile==1.1
e3nn==0.5.8
omegaconf==2.3.0
PyYAML==6.0.2
pydantic==2.11.9
addict==2.4.0
matplotlib==3.9.0
ray==2.43.0
deepspeed==0.17.6
loguru==0.7.3
tqdm==4.66.5
wandb==0.22.0
uvicorn==0.38.0
Requests==2.32.5
fastapi==0.121.2
remote_pdb==2.1.0
psutil==6.0.0
evo==1.34.0
einops==0.8.1
cloudpickle==3.1.1
gsplat==1.5.3
pycolmap==3.13.0